feat: Inject version at build time
This commit introduces a change to the way the version is handled in the application. The version is now injected at build time, both in the `Makefile` and in the GitHub workflow.

This ensures that the version printed by the `--version` command is always the correct one, and that the releases are created with the correct version.